Create README.md
In this repository we will have the codes for fatigue detection and person registration and identification and GUI. This will help us to identify the person and his drowsiness and fatigue state. If drowsiness is detected we will have an alert sound and it will remain unchanged if no drowsiness is detected.